This commit did these things: - Fixed a syntax error from CI script - Remove old readme. - Update .gitignore, add more file to block. And: 1. Created Per-CPU data structures include/kernel/percpu.h - Defines per-CPU data structures to store:
kernel_rsp: kernel stack pointer

user_rsp: user stack pointer

src/kernel/percpu.c - Implements initialization of per-CPU data and sets the GS base MSR to point to the per-CPU data structures

2. Rewrote syscall_entry.S
arch/x86_64/cpu/syscall_entry.S - The full SYSCALL entry point now includes:

swapgs - switches GS base from user mode to kernel mode (per-CPU data)

stack switch - saves user RSP to per-CPU data, loads kernel RSP

register save - saves all necessary registers onto the kernel stack

argument passing - sets up parameters correctly according to the System V ABI calling convention

call syscall_handler - executes the actual system call handling

register restore - restores all registers from the kernel stack

restore user RSP - restores the user stack pointer from per-CPU data

swapgs - switches GS base back to user mode

sysretq - returns to user mode

3. Updated syscall.c
arch/x86_64/cpu/syscall.c - In enable_syscall_mechanism():

initialize per-CPU data structures

set kernel stack pointer to per-CPU data

+++ include/kernel/mm.h
@@ -111,7 +111,7 @@ void mm_free_pages(void *pages, uint32_t count);
  * mm_map_page — insert a vaddr→paddr mapping into the current page tables.
  * flags: combination of MM_FLAG_* constants.
  */
-int mm_map_page(uint64_t vaddr, uint32_t paddr, uint32_t flags);
+int mm_map_page(uint64_t vaddr, uint64_t paddr, uint32_t flags);
 int mm_unmap_page(uint64_t vaddr);
 
 /*
